+++ conf/modules/actuators_dshot.xml
@@ -8,6 +8,11 @@
       Beware that servo output from the same timer cannot mix PWM and DSHOT.
       It might be required to disable by hand some PWM output to avoid conflicts when they are activated by default on a board.
       Currently only implemented over ChibiOS.
+
+      Configure DSHOT_BIDIR to TRUE to enable bidirectionnal DSHOT.
+      It overrides the rpm values from the UART dshot telemetry.
+      UART dshot telemetry is still usefull for current and voltage telemetry.
+
     </description>
     <configure name="DSHOT_BIDIR" value="FALSE" description="Bidirectionnal DSHOT for fast rpm feedback"/>
     <define name="DSHOT_SPEED" value="600" description="DSHOT speed (150,300,600,1200)"/>
@@ -27,7 +32,7 @@
     <file_arch name="esc_dshot.c" cond="ifeq ($(RTOS),chibios)"/>
     <file_arch name="dshot_rpmCapture.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
     <file_arch name="dshot_erps.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
-    <file_arch name="inputCapture.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
+    <file_arch name="input_capture_arch.c" dir="mcu_periph" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
     <file_arch name="timerDmaCache.c" cond="ifeq ($(DSHOT_BIDIR),TRUE)"/>
     <file_arch name="hal_stm32_dma.c" dir="mcu_periph" cond="ifeq ($(RTOS),chibios)"/>
     <test/>
