+++ sw/airborne/arch/chibios/modules/actuators/dshot_rpmCapture.h
@@ -2,7 +2,7 @@
 
 #include <ch.h>
 #include <hal.h>
-#include "inputCapture.h"
+#include "mcu_periph/input_capture_arch.h"
 #include "mcu_periph/hal_stm32_dma.h"
 #include "timerDmaCache.h"
 #include "modules/actuators/esc_dshot_config.h"
